Allow adjust! on delivery step
If an order has no total the default Solidus state machine will
skip the payment step. This means that the adjust! method won't
call `super` until the user completes the order making the created
tax adjustments unfinalized. For the most part this won't matter
since if the order total is 0 the calculated tax will most likely
be 0 as well but keeping structure and consistency to Solidus'
state_machine seems like idea.