Avoid breaking Calculator when `avalara_response` is not a hash
I ran on a scenario where the `avalara_response` would not be `nil` but not be a Hash either when the request failed - causing a cryptic `undefined method :each for nil class` up the stack. It was related to imported Orders that tried to calculate tax with Avalara.

In any case, this is just a naive fix, but it should hopefully save some headache from people that encounter something similar in the future.